Filter-Search Buttons added.
When you click on of the 8 btns the background changes to red, so its active. In the hidden Input fields the states of the buttons are saved.
So we can use the informations for the case Search.

The idea is, if you want to search something, you can filter some informations out. For Example you click the btn Gender. Now if you type Male or M in the Searchbar, only the Males are getting showed in the crud.