Started with Invoice Website
Structurally like contact.php.
Started with Modal;
Dropdown Selection for all Customer (First all Organizations sorted from A to Z || Then all Persons sorted from A to Z)
Street and zip code, etc. are displayed in the modal, so if there is a possibility of people with the same name, we can look at the street and distinguish the right person.
Clicking outside the form-container (almost outside the modal) will close or not show the customer details.