ckEditor/Leistung Inputfields load the contents of the edited invoice
Unfortunately, you cannot use the editor.setData() later because the created & saved editors (firstEditor[0] & editorArray[]) are defined as undefined for an unfounded reason.

But we are using these Editor Arrays to check, if a ckEditor Inputfield is empty or not. Its working even they are called undefined with the help of console.log(firstEditor[0]);